requirements.txt
fastapi[standard]
beanie
pydantic-settings
passlib[bcrypt]
cloudinary
motor>=3.0.0
google-genai
pyjwt